Refactor file paths and import statements in SearchPage and ChurchMap components
